qr download functionality added

+++ frontend/src/Components/Signup/Signup.js
@@ -13,29 +13,37 @@ function Signup() {
   const [confirm_password, setConfirm_passwrd] = useState('')
   const [bio, setBio] = useState('')
 
+
   const handleSignUp = async ()=>{
-    try{
-      const response = await axios.post(
-        BASE_URL+"create/",
-        { "username": username, "password":password,"bio":bio }
-      )
-      localStorage.setItem("token",response.data['token'])
-      localStorage.setItem("user_id",response.data['user_id'])
-      navigate('/')
+    if(username != '' && password != '' && confirm_password != '' && bio != ''){
+      if(password == confirm_password){
+        try{
+          const response = await axios.post(
+            BASE_URL+"create/",
+            { "username": username, "password":password,"bio":bio }
+          )
+          localStorage.setItem("token",response.data['token'])
+          localStorage.setItem("user_id",response.data['user_id'])
+          navigate('/')
+        }
+        catch(error){
+          alert(error.response.data['error'])
+        }
+      }
+  
+      else{
+        alert("Password field doesn't match!!")
+      }
     }
-    catch(error){
-      alert(error.response.data['error'])
+    else{
+      alert("Fill out every fields!!")
     }
-    
-   
-    
-    
    
   }
 
   const handleLogin = ()=>{
     console.log("Method called")
-    navigate('/')
+    navigate('/login')
     
   }
 
@@ -47,12 +55,13 @@ function Signup() {
         <form  className="s-form">
           <input type="text" value={username} className='s-input' 
                 placeholder='Username' onInput={e => setUsername(e.target.value)}/> <br/>
+          {}
           <input type="password" value={password} className='s-input' 
                 placeholder='Password' onInput={e => setPassword(e.target.value)}/> <br/>
           <input type="password" value={confirm_password} className='s-input' 
                 placeholder='Confirm Password' onInput={e => setConfirm_passwrd(e.target.value)}/> <br/>
-          <input type="text" value={bio} className='s-input' 
-                placeholder='Bio' onInput={e => setBio(e.target.value)}/> <br/>
+          <textarea value={bio} className='s-input' 
+                placeholder='Bio' onInput={e => setBio(e.target.value)}></textarea> <br/>
           <button type="button" onClick={handleSignUp} className='Signup-btn' >SIGNUP</button>
           <h6 onClick={handleLogin}>Login</h6>
         </form>
